requirements_v3.10.txt
gunicorn==20.1.0
Django==4.1
django-crispy-forms==1.14.0
crispy-bootstrap4
django-utils-six==2.0
djangorestframework==3.13.1
mysqlclient==2.1.1
PyMySQL==1.0.2
python-dotenv==0.21.1
pandas==1.5.3
SQLAlchemy==1.4.46
# matplotlib==3.7.0
sshtunnel==0.4.0
plotly==5.13.0
statsmodels==0.13.5
yfinance==0.2.4
# requests==2.31.0
requests-cache==1.0.1
# requests-ratelimiter==0.4.0
# pyrate-limiter==2.10.0
kafka-python==2.0.2